modules/requirements.txt
chainlit
ollama==0.4.7
asyncpg
boto3
requests
opencv-python
supervision
pillow
rich
